+++ webapp/publisher/snaps/views.py
@@ -44,6 +44,7 @@ from webapp.publisher.snaps.builds import (
     map_build_and_upload_states,
 )
 from werkzeug.exceptions import Unauthorized
+from webapp.publisher.snaps.build_views import get_validate_repo
 
 
 publisher_snaps = flask.Blueprint(
@@ -101,6 +102,11 @@ def _handle_error_list(errors):
 
     error_messages = ", ".join(codes)
     return flask.abort(502, error_messages)
+
+
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds/validate-repo", view_func=get_validate_repo
+)
 
 
 @publisher_snaps.route("/account/snaps")
@@ -1464,34 +1470,3 @@ def post_trigger_build(snap_name):
     return flask.redirect(
         flask.url_for(".get_snap_builds", snap_name=snap_name)
     )
-
-
-@publisher_snaps.route("/<snap_name>/builds/validate-repo", methods=["GET"])
-@login_required
-def get_validate_repo(snap_name):
-    try:
-        details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    github = GitHubAPI(flask.session.get("github_auth_secret"))
-    owner, repo = flask.request.args.get("repo").split("/")
-    response = {"success": True}
-
-    # The yaml is not present
-    if not github.get_snapcraft_yaml_location(owner, repo):
-        response["success"] = False
-        response["error"] = "MISSING_YAML_FILE"
-    # The property name inside the yaml file doesn't match the snap
-    elif not github.check_snapcraft_yaml_name(
-        owner, repo, details["snap_name"]
-    ):
-        response["success"] = False
-        response["error"] = "SNAP_NAME_DOES_NOT_MATCH"
-
-    return flask.jsonify(response)
